+++ NEWS.md
@@ -18,6 +18,7 @@
 * `object_name_linter()` now excludes special R hook functions such as `.onLoad` (#500, #614, @AshesITR)
 * `equals_na_linter()` now lints `x != NA` and `NA == x`, and skips usages in comments (#545, @michaelchirico)
 * Malformed Rmd files now cause a lint instead of an error (#571, #575, @AshesITR)
+* `spaces_inside_linter` ignores spaces preceding trailing comments (#636, @michaelchirico)
 
 # lintr 2.0.1